FIX: テーブルとカラムの修正
