+++ model/schema/schema.go
@@ -2,6 +2,7 @@ package schema
 
 import (
 	"gorm.io/gorm"
+	"time"
 )
 
 type Role string
@@ -13,15 +14,33 @@ const (
 
 type User struct {
 	gorm.Model
-	Name string `gorm:"type:varchar(255);not null"`
-	Email    string `gorm:"type:varchar(255);uniqueIndex;not null"`
-	Password string `gorm:"type:varchar(255);not null"`
-	Role Role `gorm:"type:ENUM('USER', 'ADMIN');default:'USER';not null"`
-	Books []Book
+	Name 								string 	`gorm:"type:varchar(255);not null"`
+	Email    						string 	`gorm:"type:varchar(255);uniqueIndex;not null"`
+	Password 						string 	`gorm:"type:varchar(255);not null"`
+	Role 								Role 		`gorm:"type:ENUM('USER', 'ADMIN');default:'USER';not null"`
+	Books 							[]Book
+	BorrowedBooks 			[]BorrowedBook
+	BorrowingWishLists 	[]BorrowingWishList
 }
 
 type Book struct {
 	gorm.Model
-	Title  string `gorm:"type:varchar(255);not null"`
-	UserID uint
+	UserId		uint
+	Title  		string `gorm:"type:varchar(255);not null"`
+	ImageUrl  string `gorm:"type:varchar(255);not null"`
+	Loanable  bool   `gorm:"not null"`
+}
+
+type BorrowedBook struct {
+	gorm.Model
+	UserID        uint      `gorm:"not null"`
+	BookID        uint      `gorm:"not null"`
+	CheckoutDate  time.Time `gorm:"not null"`
+	ReturnDueDate time.Time `gorm:"not null"`
+}
+
+type BorrowingWishList struct {
+	gorm.Model
+	UserID    uint      `gorm:"not null"`
+	BookID    uint      `gorm:"not null"`
 }
